+++ addon/services/syncer.js
@@ -1,9 +1,10 @@
 import Ember from 'ember';
 import Builder from '../query/builder';
+import { SimplePredicate } from '../query/predicate';
 import { reloadLocalRecords, syncDownRelatedRecords } from '../utils/reload-local-records';
 import isModelInstance from '../utils/is-model-instance';
 
-const { RSVP } = Ember;
+const { RSVP, isNone, isArray, getOwner, get } = Ember;
 
 /**
   @class Syncer
@@ -33,35 +34,36 @@ export default Ember.Service.extend({
   init: function() {
     let _this = this;
 
-    let localStore = Ember.getOwner(this).lookup('store:local');
+    let localStore = getOwner(this).lookup('store:local');
 
     _this.set('offlineStore', localStore);
   },
 
   /**
-   * TODO:
-   * Save all records in the store into local store.
+   * Save specified records into local store (IndexedDB).
    *
    * @method syncDown
    * @public
    * @param {String|DS.Model|Array} descriptor typeName, record, records.
    * @param {Boolean} [reload] If set to true then syncer perform remote reload for data, otherwise data will get from the store.
    * @param {String} [projectionName] Name of projection for remote reload of data. If not set then all properties of record, except navigation properties, will be read.
-   * @param {Query.QueryObject} [queryObject] QueryObject instance to make query if descriptor is a typeName.
+   * @param {Object} [params] Additional parameters for syncing down.
+   * @param {Query.QueryObject} [params.queryObject] QueryObject instance to make query if descriptor is a typeName.
+   * @param {Boolean} [params.unloadSyncedRecords] If set to true then synced records will be unloaded from online store.
    * @return {Promie}
    */
-  syncDown: function(descriptor, reload, projectionName, queryObject) {
+  syncDown: function(descriptor, reload, projectionName, params) {
     let _this = this;
 
     if (typeof descriptor === 'string') {
-      return reloadLocalRecords.call(this, descriptor, reload, projectionName, queryObject);
+      return reloadLocalRecords.call(this, descriptor, reload, projectionName, params);
 
     } else if (isModelInstance(descriptor)) {
-      return _this._syncDownRecord(descriptor, reload, projectionName);
+      return _this._syncDownRecord(descriptor, reload, projectionName, params);
 
-    } else if (Ember.isArray(descriptor)) {
+    } else if (isArray(descriptor)) {
       let updatedRecords = descriptor.map(function(record) {
-        return _this._syncDownRecord(record, reload, projectionName);
+        return _this._syncDownRecord(record, reload, projectionName, params);
       });
       return RSVP.all(updatedRecords);
 
@@ -87,18 +89,34 @@ export default Ember.Service.extend({
    * @param {DS.Model} record Record to save in local store.
    * @param {Boolean} [reload] If set to true then syncer perform remote reload for data, otherwise data will get from the store.
    * @param {String} [projectionName] Name of projection for remote reload of data. If not set then all properties of record, except navigation properties, will be read.
+   * @param {Object} [params] Additional parameters for syncing down.
+   * @param {Query.QueryObject} [params.queryObject] QueryObject instance to make query if descriptor is a typeName.
+   * @param {Boolean} [params.unloadSyncedRecords] If set to true then synced records will be unloaded from online store.
    * @private
    */
-  _syncDownRecord: function(record, reload, projectionName) {
+  _syncDownRecord: function(record, reload, projectionName/*, params*/) {
+    var _this = this;
+
     function saveRecordToLocalStore(store, record, projectionName) {
-      let dexieService = Ember.getOwner(this).lookup('service:dexie');
+      let dexieService = getOwner(_this).lookup('service:dexie');
       dexieService.set('queueSyncDownWorksCount', dexieService.get('queueSyncDownWorksCount') + 1);
       let modelName = record.constructor.modelName;
       let modelType = store.modelFor(modelName);
-      let projection = Ember.isNone(projectionName) ? null : modelType.projections[projectionName];
+      let projection = isNone(projectionName) ? null : modelType.projections[projectionName];
       let localStore = this.get('offlineStore');
       let localAdapter = localStore.adapterFor(modelName);
       let snapshot = record._createSnapshot();
+      let unloadRecordFromStore = () => {
+
+        // TODO: Uncomment this after fix bug with load unloaded models.
+        // if (params && params.unloadSyncedRecords) {
+        //   if (store.get('onlineStore')) {
+        //     store.get('onlineStore').unloadRecord(record);
+        //   } else {
+        //     store.unloadRecord(record);
+        //   }
+        // }
+      };
 
       if (record.get('isDeleted')) {
         return localAdapter.deleteRecord(localStore, snapshot.type, snapshot).then(() => {
@@ -106,28 +124,37 @@ export default Ember.Service.extend({
         }).catch((reason) => {
           dexieService.set('queueSyncDownWorksCount', dexieService.get('queueSyncDownWorksCount') - 1);
           Ember.Logger.error(reason);
-        });
+        }).finally(unloadRecordFromStore);
       } else {
         return localAdapter.updateOrCreate(localStore, snapshot.type, snapshot).then(function() {
           dexieService.set('queueSyncDownWorksCount', dexieService.get('queueSyncDownWorksCount') - 1);
-          return projection ? syncDownRelatedRecords(store, record, localAdapter, localStore, projection) : RSVP.resolve();
+          let offlineGlobals = Ember.getOwner(_this).lookup('service:offline-globals');
+          if (projection || (!projection && offlineGlobals.get('allowSyncDownRelatedRecordsWithoutProjection'))) {
+            return syncDownRelatedRecords.call(_this, store, record, localAdapter, localStore, projection);
+          } else {
+            Ember.Logger.warn('It does not allow to sync down related records without specified projection. ' +
+              'Please specify option "allowSyncDownRelatedRecordsWithoutProjection" in environment.js');
+            return RSVP.resolve();
+          }
         }).catch((reason) => {
           dexieService.set('queueSyncDownWorksCount', dexieService.get('queueSyncDownWorksCount') - 1);
           Ember.Logger.error(reason);
-        });
+        }).finally(unloadRecordFromStore);
       }
     }
 
-    var _this = this;
-    var store = Ember.getOwner(this).lookup('service:store');
+    var store = getOwner(this).lookup('service:store');
     if (reload) {
       let modelName = record.constructor.modelName;
       let options = {
         reload: true,
         useOnlineStore: true
       };
-      options = Ember.isNone(projectionName) ? options : Ember.$.extend(true, options, { projection: projectionName });
+      options = isNone(projectionName) ? options : Ember.$.extend(true, options, { projection: projectionName });
       return store.findRecord(modelName, record.id, options).then(function(reloadedRecord) {
+
+        // TODO: Uncomment this after fix bug with load unloaded models.
+        // store.get('onlineStore').unloadRecord(modelName);
         return saveRecordToLocalStore.call(_this, store, reloadedRecord, projectionName);
       });
     } else {
@@ -138,16 +165,18 @@ export default Ember.Service.extend({
   /**
   */
   syncUp(continueOnError) {
-    let _this = this;
-    let store = Ember.getOwner(this).lookup('service:store');
-    return _this.get('offlineStore').query('i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-audit-entity', {
-      // TODO: Needs sort by `operationTime`.
-      // TODO: After inject query language support, add filter `executionResult` by `Unexecuted` or `Failed`.
-      executionResult: 'Unexecuted',
-    }).then((jobs) => {
-      // TODO: Delete `sortBy` after sort by `operationTime`.
-      jobs = jobs.sortBy('operationTime');
-      return _this._runJobs(store, jobs, continueOnError);
+    let dexieService = getOwner(this).lookup('service:dexie');
+    let store = getOwner(this).lookup('service:store');
+    let modelName = 'i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-audit-entity';
+    let predicate = new SimplePredicate('executionResult', 'eq', 'Unexecuted')
+      .or(new SimplePredicate('executionResult', 'eq', 'Failed'));
+    let builder = new Builder(store, modelName)
+      .selectByProjection('AuditEntityE')
+      .orderBy('operationTime')
+      .where(predicate);
+    return this.get('offlineStore').query(modelName, builder.build()).then((jobs) => {
+      dexieService.set('queueSyncUpTotalWorksCount', jobs.get('length'));
+      return this._runJobs(store, jobs, continueOnError);
     });
   },
 
@@ -170,6 +199,8 @@ export default Ember.Service.extend({
   */
   _runJobs(store, jobs, continueOnError, jobCount) {
     let _this = this;
+    let dexieService = getOwner(_this).lookup('service:dexie');
+    dexieService.set('queueSyncUpWorksCount', jobs.get('length'));
     let job = jobs.shiftObject();
     let executedJob = jobCount || 0;
     return new RSVP.Promise((resolve, reject) => {
@@ -186,6 +217,7 @@ export default Ember.Service.extend({
           });
         });
       } else {
+        dexieService.set('queueSyncUpWorksCount', 0);
         resolve(executedJob);
       }
     });
@@ -194,9 +226,11 @@ export default Ember.Service.extend({
   /**
   */
   _endJob(job) {
+    let dexieService = getOwner(this).lookup('service:dexie');
     return new RSVP.Promise((resolve, reject) => {
       if (job.get('executionResult') === 'Выполнено') {
         RSVP.all(job.get('auditFields').map(field => field.destroyRecord())).then(() => {
+          dexieService.set('queueSyncUpCurrentModelName', null);
           resolve(job.destroyRecord());
         });
       } else {
@@ -208,6 +242,9 @@ export default Ember.Service.extend({
   /**
   */
   _runJob(store, job) {
+    let dexieService = getOwner(this).lookup('service:dexie');
+    dexieService.set('queueSyncUpCurrentModelName', job.get('objectType.name'));
+
     switch (job.get('operationType')) {
       case 'INSERT': return this._runCreatingJob(store, job);
       case 'UPDATE': return this._runUpdatingJob(store, job);
@@ -219,19 +256,38 @@ export default Ember.Service.extend({
   /**
   */
   _runCreatingJob(store, job) {
-    let attributes = {
-      id: job.get('objectPrimaryKey'),
-    };
-    job.get('auditFields').forEach((field) => {
-      attributes[field.get('field')] = field.get('newValue');
-    });
-    return store.createRecord(job.get('objectType.name'), attributes, true).save().then(() => {
-      job.set('executionResult', 'Выполнено');
-      return job.save();
-    }).catch((/*reason*/) => {
-      // TODO: Resolve conflicts here.
-      job.set('executionResult', 'Ошибка');
-      return job.save();
+    let record = store.peekRecord(job.get('objectType.name'), job.get('objectPrimaryKey'));
+
+    // TODO: Uncomment this after fix bug with load unloaded models.
+    // if (record) {
+    //   record.rollbackAll();
+    //   store.unloadRecord(record);
+    // }
+
+    if (!record)
+    {
+      record = store.createRecord(job.get('objectType.name'), { id: job.get('objectPrimaryKey') });
+    }
+
+    record.set('isSyncingUp', true);
+    record.set('isCreatedDuringSyncUp', true);
+
+    return this._changesForRecord(store, job).then((changes) => {
+      record.setProperties(changes);
+      return record.save().then(() => {
+        job.set('executionResult', 'Выполнено');
+        return job.save();
+      }).catch((reason) => {
+
+        // TODO: Resolve conflicts here.
+        job.set('executionResult', 'Ошибка');
+        Ember.Logger.error(`Sync up model '${job.get('objectType.name')}' creating job error`, reason, record);
+        return job.save();
+      }).finally(() => {
+        if (record) {
+          record.set('isSyncingUp', false);
+        }
+      });
     });
   },
 
@@ -241,16 +297,24 @@ export default Ember.Service.extend({
     let query = this._createQuery(store, job);
     return store.queryRecord(query.modelName, query).then((record) => {
       if (record) {
-        job.get('auditFields').forEach((field) => {
-          record.set(field.get('field'), field.get('newValue'));
-        });
-        return record.save().then(() => {
-          job.set('executionResult', 'Выполнено');
-          return job.save();
-        }).catch((/*reason*/) => {
-          // TODO: Resolve conflicts here.
-          job.set('executionResult', 'Ошибка');
-          return job.save();
+        record.set('isSyncingUp', true);
+        return this._changesForRecord(store, job).then((changes) => {
+          record.setProperties(changes);
+          return record.save().then(() => {
+            record.set('isUpdatedDuringSyncUp', true);
+            job.set('executionResult', 'Выполнено');
+            return job.save();
+          }).catch((reason) => {
+
+            // TODO: Resolve conflicts here.
+            job.set('executionResult', 'Ошибка');
+            Ember.Logger.error(`Sync up model '${query.modelName}' updating job error`, reason, record);
+            return job.save();
+          }).finally(() => {
+            if (record) {
+              record.set('isSyncingUp', false);
+            }
+          });
         });
       } else {
         throw new Error('No record is found on server.');
@@ -264,13 +328,21 @@ export default Ember.Service.extend({
     let query = this._createQuery(store, job);
     return store.queryRecord(query.modelName, query).then((record) => {
       if (record) {
+        record.set('isSyncingUp', true);
         return record.destroyRecord().then(() => {
+          record.set('isDestroyedDuringSyncUp', true);
           job.set('executionResult', 'Выполнено');
           return job.save();
-        }).catch((/*reason*/) => {
+        }).catch((reason) => {
+
           // TODO: Resolve conflicts here.
           job.set('executionResult', 'Ошибка');
+          Ember.Logger.error(`Sync up model '${query.modelName}' removing job error`, reason, record);
           return job.save();
+        }).finally(() => {
+          if (record) {
+            record.set('isSyncingUp', false);
+          }
         });
       } else {
         throw new Error('No record is found on server.');
@@ -350,7 +422,7 @@ export default Ember.Service.extend({
   */
   _auditDataFromRecord(record) {
     let _this = this;
-    let userService = Ember.getOwner(_this).lookup('service:user');
+    let userService = getOwner(_this).lookup('service:user');
     let operationType = _this._getOperationType(record.get('dirtyType'));
     return userService.getCurrentUser().then(currentUser => _this._getObjectType(record._createSnapshot().modelName).then(objectType => ({
           objectPrimaryKey: record.get('id'),
@@ -398,6 +470,49 @@ export default Ember.Service.extend({
 
   /**
   */
+  _changesForRecord(store, job) {
+    return new RSVP.Promise((resolve, reject) => {
+      let changes = {};
+      let promises = [];
+      let attributes = get(store.modelFor(job.get('objectType.name')), 'attributes');
+      job.get('auditFields').forEach((auditField) => {
+        let descriptorField = auditField.get('field').split('@');
+        let field = descriptorField.shift();
+        let type = descriptorField.shift();
+        if (type) {
+          let builder = new Builder(store, type).byId(auditField.get('newValue'));
+          promises.push(store.queryRecord(type, builder.build()).then((relationship) => {
+            changes[field] = relationship;
+          }));
+        } else {
+          let value = auditField.get('newValue');
+          switch (attributes.get(field).type) {
+            case 'boolean':
+              changes[field] = value === null ? null : !!value;
+              break;
+
+            case 'number':
+              changes[field] = value === null ? null : +value;
+              break;
+
+            case 'date':
+              changes[field] = value === null ? null : new Date(value);
+              break;
+
+            default:
+              changes[field] = value;
+          }
+        }
+      });
+
+      RSVP.all(promises).then(() => {
+        resolve(changes);
+      }, reject);
+    });
+  },
+
+  /**
+  */
   _changesFromRecord(record) {
     let changes = {};
     if (this.get('auditEnabled')) {
@@ -413,6 +528,24 @@ export default Ember.Service.extend({
         }
       });
     }
+
+    let changedRelationships = record.changedBelongsTo();
+    let snapshot = record._createSnapshot();
+    record.eachRelationship((name, descriptor) => {
+      let changedRelationship = changedRelationships[name];
+      if (changedRelationship && descriptor.kind === 'belongsTo') {
+        let relationshipType = descriptor.type;
+        if (descriptor.options && descriptor.options.polymorphic) {
+          let belongsTo = snapshot.belongsTo(descriptor.key);
+          relationshipType = belongsTo.modelName;
+        }
+
+        changes[`${name}@${relationshipType}`] = [
+          changedRelationship[0] && changedRelationship[0].get('id'),
+          changedRelationship[1] && changedRelationship[1].get('id'),
+        ];
+      }
+    });
 
     return changes;
   },