+++ addon/stores/local-store.js
@@ -169,7 +169,18 @@ export default DS.Store.extend({
   query: function(modelName, query) {
     Ember.Logger.debug(`Flexberry Local Store::query ${modelName}`, query);
 
-    return this._super(...arguments);
+    let promise = this._super(...arguments);
+    return new Ember.RSVP.Promise((resolve, reject) => {
+      promise.then((results) => {
+        if (results && Ember.isArray(results)) {
+          results.forEach((result) => {
+            result.didLoad();
+          });
+        }
+
+        resolve(results);
+      }, reject);
+    });
   },
 
   /**
@@ -190,6 +201,15 @@ export default DS.Store.extend({
   queryRecord: function(modelName, query) {
     Ember.Logger.debug(`Flexberry Local Store::queryRecord ${modelName}`, query);
 
-    return this._super(...arguments);
+    let promise = this._super(...arguments);
+    return new Ember.RSVP.Promise((resolve, reject) => {
+      promise.then((result) => {
+        if (result) {
+          result.didLoad();
+        }
+
+        resolve(result);
+      }, reject);
+    });
   },
 });