+++ CHANGELOG.md
@@ -2,6 +2,104 @@
 
 ## [Unreleased]
 
+## [0.7.1-beta.2] - 2016-10-31
+### Added
+- Now `dexie` service can work with multiple DBs.
+- Now hashes received via `dexie` service have `loadRelationships` method that can be used for replacing ids with hashes when reading synchronous relationships.
+- Some parameters checks for `builder` and `information` classes.
+
+### Fixed
+- Loading relationships in offline mode moved "under" offline adapter layer, i.e. offline adapter now get "full object" at once with embedded objects for embedded relationships. So now fixed:
+    - `indexeddb-adapter` apply filters after loading relationships. This allows to filter data by master and detail model's proprties.
+    - There is no relationships that "disappear" unexpectedly after reading data.
+    - Improved performance for reading data in offline mode (but for some cases improving performance is still required).
+
+## [0.7.1-beta.1] - 2016-10-28
+### Fixed
+- Filter by boolean value in offline mode, if restriction was passed via complex predicate.
+
+## [0.7.1-beta.0] - 2016-10-27
+### Fixed
+- `syncer` service now makes casting for `boolean` type and corrected casting for `null` value of all types while performing sync up.
+- Serializing attributes of `boolean`, `number` and `decimal` types when saving offline.
+- Now it is able to save and read `null` value in `boolean` attributes in offline mode.
+
+## [0.6.2-beta.30] - 2016-10-21
+### Added
+- Base model properties:
+    - `isSyncingUp` - true if model is syncing up.
+    - `isCreatedDuringSyncUp` - true if model is created during sync up process.
+    - `isUpdatedDuringSyncUp` - true if model is updated last time during sync up process.
+    - `isDestroyedDuringSyncUp` - true if model is destroyed during sync up process.
+
+### Fixed
+- Now we are saving information about model type of polymorphic belongsTo relationships into local store. Setting proper types of polymorphic belongsTo relationships should come in next versions of addon.
+
+## [0.6.2-beta.29] - 2016-10-20
+### Fixed
+- Saving of wrong relationship type for polymorphic belongsTo relationships in offline audit objects.
+
+## [0.6.2-beta.28] - 2016-10-19
+### Added
+- Saving of base models logic for offline (for create/update/delete operations). Now it is not using transactions (will be added in next versions).
+
+### Changed
+- Disable models unloads after sync down and sync up operations in `syncer` service.
+
+## [0.6.2-beta.27] - 2016-10-18
+### Fixed
+- Some internal bugs in `syncDown` method of `syncer` service.
+
+## [0.6.2-beta.26] - 2016-10-18
+### Fixed
+- If `queryRecord` method did not find any record, it still tries to call `didLoad`.
+
+## [0.6.2-beta.25] - 2016-10-17
+### Added
+- `allowSyncDownRelatedRecordsWithoutProjection` option for offline settings. If true then record will be synced down completely, i.e. including with all related records with arbitrary nesting. Otherwise only requested record will be synced down without related records.
+
+## [0.6.2-beta.24] - 2016-10-17
+### Fixed
+- When rollback `belongsTo` relationships, if relationship was reset to null then rollback relationship attempts to apply for null.
+- When rollback `belongsTo` relationships, there was no rollback for canonical state of relationship.
+- If model is not support `rollbackBelongsTo` mehod then it should not be called.
+
+### Known issues
+- Event `didLoad` fired for model instance only when model loads first time. We added manual triggering of this event to save state, because model instance can be loaded in another projection.
+
+## [0.6.2-beta.23] - 2016-10-16
+### Fixed
+- Small optimization of performance for sync down operation.
+- `syncer` service waited for loading only one last relationship when restoring changes for record while performing sync up.
+- `syncer` service was not considered types of attributes when restoring values while performing sync up, now it makes casting for `date` and `number` types.
+-  Audit fields was not filled because `save` method of `offline-model` mixin for offline run first (before `save` method of `audit-model` mixin).
+
+## [0.6.2-beta.22] - 2016-10-15
+### Added
+- `syncer` service now saves changed relationships for sync up process.
+
+### Changed
+- `syncer` service now selects jobs for sync up using query language. This allows select at once 'Unexecuted' and 'Failed' jobs and sort them.
+
+## [0.6.2-beta.21] - 2016-10-15
+### Changed
+- Now we don't replace entire record when sync down. Instead we save only changed properties which are specified in projection for existing records in local store.
+
+### Fixed
+- Update IndexedDB database failed if database version in offline schema was specified as string.
+- Sorting in `indexeddb-adapter` is now applying before skip and top.
+
+## [0.6.2-beta.20] - 2016-10-14
+### Added
+- Ability to specify option for unloading records from online store when sync down. It's topical for sync down big data. Please add `{ unloadSyncedRecords: true }` as fourth parameter in `syncDown` method calls where needed.
+
+### Changed
+- Search by string (by all fields in lists) is now case insensitive in offline mode.
+
+### Fixed
+- Not quite adequate implementation of `updateRecord` method of `offline` adapter.
+- Search by empty fields in `js-adapter`.
+
 ## [0.6.2-beta.19] - 2016-10-12
 ### Added
 - `timeout` property to `odata` adapter.